7l: long -> int32

+++ src/cmd/7l/7.out.h
@@ -511,8 +511,8 @@ enum
 typedef	struct	Ieee	Ieee;
 struct	Ieee
 {
-	long	l;	/* contains ls-man	0xffffffff */
-	long	h;	/* contains sign	0x80000000
+	int32	l;	/* contains ls-man	0xffffffff */
+	int32	h;	/* contains sign	0x80000000
 				    exp		0x7ff00000
 				    ms-man	0x000fffff */
 };
